% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{cor.cn}
\alias{cor.cn}
\title{correct copynumber by gc percent and mappability}
\usage{
cor.cn(bed, map.min = 0.9, samplesize = 50000, verbose = T, gc.cor = T,
  map.cor = T)
}
\arguments{
\item{bed}{A data.table}
}
\value{
A data.table
}
\description{
correct copynumber by gc percent and mappability
}

